Time: 16 ms (68.81%), Space: 18 MB (34.26%) - LeetHub

+++ 0003-longest-substring-without-repeating-characters/0003-longest-substring-without-repeating-characters.py
@@ -1,16 +1,18 @@
 class Solution:
     def lengthOfLongestSubstring(self, s: str) -> int:
+        if not s:
+            return 0
+
         mapper = set()
         max_length = 0
         start = 0
-        
+
         for end, c in enumerate(s):
-            while c in mapper:
+            while s[end] in mapper:
                 mapper.remove(s[start])
                 start += 1
             
-            mapper.add(c)
-
-            max_length = max(len(mapper), max_length)
+            mapper.add(s[end])
+            max_length = max(max_length, len(mapper))
 
         return max_length
